chore: refactor seperate render and layout
