+++ src/routes/scripts/layoutPack.ts
@@ -1,0 +1,202 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+// TODO: remove above line
+const FACTOR = 100;
+const PADDING_INNER_PACK = 2;
+const PADDING_INTER_PACK = 10; // must be more than 1
+const PADDING_X_LAYER = 100;
+import * as d3 from 'd3';
+const layers = ['Presentation Layer', 'Service Layer', 'Domain Layer', 'Data Source Layer'];
+
+export default function layoutPack(roots: any[]) {
+	const layerCountProto: any = layers.reduce(
+		(obj: { [key: string]: number }, item: string) => {
+			obj[item] = 0;
+			return obj;
+		},
+		{ 'Unknown Layer': 0 }
+	);
+	// COMPUTE
+	roots.forEach((root) => {
+		// 1. Compute the pack layout.
+		const pack = d3.pack<any>();
+		pack.padding(PADDING_INNER_PACK);
+
+		// // Sum and sort the data.
+		root.sum((d: any) => d.count);
+
+		// arbitrary decision: determine the max size of the root based on its value
+		const size = Math.sqrt(root.value) * FACTOR; // make it non linear
+		pack.size([size, size]);
+
+		pack(root);
+
+		// 2. find the total 'count' attribute for each layer
+		const layerCount: any = { ...layerCountProto };
+		let total = 0;
+		root.each((d: any) => {
+			const layer = d.data.layer;
+			layerCount[layer] += d.data.count;
+			total += d.data.count;
+		});
+
+		const layerPercentage = Object.keys(layerCount).reduce((acc: any, key) => {
+			acc[key] = (layerCount[key] / total) * 100;
+			return acc;
+		}, {});
+
+		// find the dominants layer, and remove non dominant layer.
+		/*
+		step
+		1. sort the layerPercentage from highest to lowest
+		2. if the second highest is more than 1.5 times the highest, the it's considered as dominant
+		3. do this until non dominan is found, then truncate the rest
+		*/
+		// turn the layerPercentage into array
+		const layerPercentageArray = Object.keys(layerPercentage).map((key) => {
+			return { layer: key, percentage: layerPercentage[key] };
+		});
+		// sort the layerPercentageArray from highest to lowest
+		layerPercentageArray.sort((a, b) => b.percentage - a.percentage);
+
+		const dominantLayer = [layerPercentageArray[0]];
+		let layerIndexToCheck = 1;
+		while (layerIndexToCheck < layerPercentageArray.length) {
+			const currentLayer = layerPercentageArray[layerIndexToCheck];
+			const previousLayer = layerPercentageArray[layerIndexToCheck - 1];
+			if (currentLayer.percentage * 1.5 > previousLayer.percentage) {
+				dominantLayer.push(currentLayer);
+			} else {
+				break;
+			}
+			layerIndexToCheck++;
+		}
+
+		// attach the layerPercentage to the root
+		root.data.dominantLayer = dominantLayer;
+		// root.data.layerPercentageArray = layerPercentageArray;
+	});
+
+	// CALCULATE LAYOUT
+	// draw the layer as rectangle
+	// height is the biggest pack size
+	const biggestRootRadius = roots.reduce((acc, root) => {
+		if (acc.value < root.value) {
+			return root;
+		}
+		return acc;
+	});
+
+	// arbitrary layer height: 3 times the biggest root radius
+	const layerHeight = biggestRootRadius.r * 3;
+	// find the y for each root based on the dominantLayer
+
+	// draw each package (which is represented by a root)
+	calculateRootsLocationBasedOnDominantLayer(roots, layerHeight);
+}
+
+type Circle = {
+	x: number;
+	y: number;
+	r: number;
+};
+function doesCircleOverlap(circle1: Circle, circle2: Circle) {
+	return findMoveNeededInXForCircle2(circle1, circle2) > 0;
+}
+
+// find the x and lock the y for the circle to not overlap new x must be to the right of the circle1
+function findMoveNeededInXForCircle2(circle1: Circle, circle2: Circle) {
+	const targetDistanceInX = Math.sqrt((circle1.r + circle2.r) ** 2 - (circle1.y - circle2.y) ** 2);
+	if (Number.isNaN(targetDistanceInX)) {
+		return 0;
+	}
+	const currentDistanceInX = Math.abs(circle1.x - circle2.x);
+	let moveNeeded = targetDistanceInX - currentDistanceInX;
+	// if circle2 is to the left, add 2 * currentDistanceInX
+	if (circle2.x < circle1.x) {
+		moveNeeded += 2 * currentDistanceInX;
+	}
+	return moveNeeded;
+}
+
+function calculateRootsLocationBasedOnDominantLayer(roots: any[], layerHeight: number) {
+	const previousRoots: any = [];
+	roots.forEach((root) => {
+		const dominantLayer = root.data.dominantLayer;
+		/*
+	if dominant layer > 2 ,set Y to be -100
+	if dominant layer = 2, check the two layer. if they're side to side, calculate based on the ratio of those two layer. 
+	if not, set Y to be -100
+	if dominant layer = 1, set Y the center of that layer
+*/
+		if (dominantLayer.length > 2) {
+			root.containerY = -100;
+			root.isCrossOver = true;
+		} else if (dominantLayer.length === 2) {
+			const layer1 = dominantLayer[0].layer;
+			const layer2 = dominantLayer[1].layer;
+			const layer1Index = layers.indexOf(layer1);
+			const layer2Index = layers.indexOf(layer2);
+			if (layer1Index + 1 === layer2Index || layer1Index - 1 === layer2Index) {
+				const layer1 = dominantLayer[0].layer;
+				const layer1Index = layers.indexOf(layer1);
+				const layer1Location = (layer1Index + 0.5) * layerHeight;
+				const layer1Percentage = dominantLayer[0].percentage;
+
+				const layer2 = dominantLayer[1].layer;
+				const layer2Index = layers.indexOf(layer2);
+				const layer2Location = (layer2Index + 0.5) * layerHeight;
+				const layer2Percentage = dominantLayer[1].percentage;
+
+				// calculate based on ratio
+				const distanceAdded =
+					(layer2Location - layer1Location) *
+					(layer2Percentage / (layer1Percentage + layer2Percentage));
+				root.containerY = layer1Location + distanceAdded;
+			} else {
+				root.containerY = -100;
+				root.isCrossOver = true;
+			}
+		} else if (dominantLayer.length === 1) {
+			const layer = dominantLayer[0].layer;
+			const layerIndex = layers.indexOf(layer);
+			root.containerY = (layerIndex + 0.5) * layerHeight;
+		} else {
+			root.containerY = -100;
+			root.isCrossOver = true;
+		}
+		// the coordinate is from top left.
+		root.containerY -= root.r;
+
+		// calculate the x. check if it overlaps with previous roots. if it does, move it to the right then check again
+		root.containerX = PADDING_X_LAYER;
+
+		// skip if crossover
+		if (root.isCrossOver) {
+			return;
+		}
+
+		let isOverlap = false;
+		do {
+			isOverlap = false;
+			for (let i = 0; i < previousRoots.length; i++) {
+				const previousRoot = previousRoots[i];
+				const circle1 = {
+					x: previousRoot.containerX + previousRoot.r, // center of the circle
+					y: previousRoot.containerY + previousRoot.r, // center of the circle
+					r: previousRoot.r
+				};
+				const circle2 = {
+					x: root.containerX + root.r, // center of the circle
+					y: root.containerY + root.r, // center of the circle
+					r: root.r
+				};
+				if (doesCircleOverlap(circle1, circle2)) {
+					isOverlap = true;
+					root.containerX += findMoveNeededInXForCircle2(circle1, circle2) + PADDING_INTER_PACK;
+					break;
+				}
+			}
+		} while (isOverlap);
+		previousRoots.push(root);
+	});
+}
