+++ src/routes/scripts/renderPack.ts
@@ -1,0 +1,177 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+// TODO: remove above line
+import { adjustHexColor } from '$lib';
+import * as d3 from 'd3';
+const PADDING_X_LAYER = 100;
+const layers = ['Presentation Layer', 'Service Layer', 'Domain Layer', 'Data Source Layer'];
+
+export function renderLayer(
+	canvas: d3.Selection<SVGGElement, unknown, null, undefined>,
+	roots: any[]
+) {
+	const biggestRootRadius = roots.reduce((acc, root) => {
+		if (acc.value < root.value) {
+			return root;
+		}
+		return acc;
+	});
+
+	// arbitrary layer height: 3 times the biggest root radius
+	const layerHeight = biggestRootRadius.r * 3;
+	// TODO: parameterize?
+	const colorMap: { [key: string]: string } = {
+		'Presentation Layer': '#fb8072',
+		'Service Layer': '#ffffb3',
+		'Domain Layer': '#8dd3c7',
+		'Data Source Layer': '#bebada',
+		'Unknown Layer': '#525252'
+	};
+
+	const defs = canvas.append('svg:defs');
+
+	Object.entries(colorMap).forEach(([key, value]) => {
+		const grad = defs
+			.append('svg:radialGradient')
+			.attr('gradientUnits', 'objectBoundingBox')
+			.attr('cx', '25%')
+			.attr('cy', '25%')
+			.attr('r', '100%')
+			.attr('id', 'grad' + key.split(' ').join(''));
+
+		grad.append('stop').attr('offset', '0%').style('stop-color', 'white');
+
+		grad.append('stop').attr('offset', '75%').style('stop-color', adjustHexColor(value, -0.8));
+	});
+
+	const grad = defs
+		.append('svg:radialGradient')
+		.attr('gradientUnits', 'objectBoundingBox')
+		.attr('cx', '25%')
+		.attr('cy', '25%')
+		.attr('r', '100%')
+		.attr('id', 'gradElse');
+
+	grad.append('stop').attr('offset', '0%').style('stop-color', '#666666');
+
+	grad.append('stop').attr('offset', '75%').style('stop-color', '#cccccc');
+
+	// calculate max width
+	let maxWidth = 0;
+	roots.forEach((root) => {
+		if (root.containerX + root.r * 2 > maxWidth) {
+			maxWidth = root.containerX + root.r * 2;
+		}
+	});
+	// RENDER LAYER
+	const layerWidth = maxWidth + PADDING_X_LAYER;
+	layers.forEach((layer, i) => {
+		canvas
+			.append('rect')
+			.attr('x', 0)
+			.attr('y', i * layerHeight)
+			.attr('width', layerWidth)
+			.attr('height', layerHeight)
+			.attr('fill', () => adjustHexColor(colorMap[layer] ?? '#cccccc', 0.4))
+			.attr('stroke', 'black');
+
+		canvas
+			.append('text')
+			.attr('x', 5)
+			.attr('y', i * layerHeight + 15)
+			.attr('fill', 'black')
+			.attr('text-anchor', 'start')
+			.attr('font-size', '14px')
+			.text(layer);
+	});
+}
+export function renderPack(
+	canvas: d3.Selection<SVGGElement, unknown, null, undefined>,
+	roots: any[],
+	writeDetailHover: (detail: string) => void
+) {
+	let maxWidth = 0;
+	roots.forEach((root) => {
+		if (root.containerX + root.r * 2 > maxWidth) {
+			maxWidth = root.containerX + root.r * 2;
+		}
+	});
+	// RENDER LAYER
+	const layerWidth = maxWidth + PADDING_X_LAYER;
+	// RENDER PACKS
+	roots.forEach((root, index: number) => {
+		const containerElement = canvas
+			.append('g')
+			.attr('transform', `translate(${root.containerX},${root.containerY})`)
+			.attr('id', `root-${index}`);
+
+		// add container drag
+		containerElement.call(
+			d3.drag<any, any>().on('drag', function (event) {
+				// if not cross over
+				if (!root.isCrossOver) {
+					const x = event.dx;
+					// constraint on y axis
+					const y = 0;
+
+					root.containerX += x;
+					root.containerY += y;
+					// limit left and right boundary
+					if (root.containerX < 0) {
+						root.containerX = 0;
+					}
+					if (root.containerX + root.r * 2 > layerWidth) {
+						root.containerX = layerWidth - root.r * 2;
+					}
+				} else {
+					const x = event.dx;
+					const y = event.dy;
+					root.containerX += x;
+					root.containerY += y;
+				}
+				d3.select(this).attr('transform', `translate(${root.containerX},${root.containerY})`);
+			})
+		);
+
+		root.each((d: any) => {
+			const node = containerElement
+				.append('g')
+				.attr('transform', `translate(${d.x},${d.y})`)
+				.attr('class', 'node');
+
+			node
+				.append('circle')
+				.attr('r', d.r)
+				.attr('fill', 'url(#grad' + (d.data?.layer?.split(' ')?.join('') ?? 'Else') + ')')
+				.attr('stroke', 'black');
+
+			node.on('mouseover', function () {
+				const hoverDetail = JSON.stringify(d.data, null, '\t');
+				writeDetailHover(hoverDetail);
+
+				d3.select(this).select('circle').attr('stroke-width', 2);
+			});
+
+			node.on('mouseout', function () {
+				writeDetailHover('');
+				d3.select(this).select('circle').attr('stroke-width', 1);
+			});
+
+			node.on('click', function () {});
+
+			node.on('contextmenu', (event, d) => {
+				event.preventDefault();
+				console.log('right click', d);
+			});
+			// only show text for package
+			if (d.data.kind === 'package') {
+				node
+					.append('text')
+					.attr('dy', -d.r - 4)
+					.attr('fill', 'black')
+					.attr('text-anchor', 'middle')
+					.attr('font-size', '12px')
+					.text(d.data.id);
+			}
+		});
+	});
+}
